+++ CHANGELOG.md
@@ -31,6 +31,7 @@ in 5.x, 4.x, 3.x, 2.x, 1.x and 0.x versions.
   * Improved part's material detection by ignoring edges banding and veneer material types applied on faces
   * Improved summary cells: unavailable cells in parts list summary are hatched instead of displaying "-" character
   * Improved UX : Added "escape" key to 1. close modals, 2. minimize OpenCutList dialog
+  * Improved UX : Added Double click behavior on material button of the Smart Paint Tool to edit the corresponding material
 
 * 4.1.3 (2023-02-28)